requirements.txt
absl-py==2.1.0
alembic==1.13.3
amqp==5.2.0
aniso8601==9.0.1
astunparse==1.6.3
attrs==24.2.0
bcrypt==4.2.0
beautifulsoup4==4.12.3
billiard==4.2.1
blinker==1.8.2
celery==5.4.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.0.7
flask-restx==1.3.0
Flask-SQLAlchemy==3.1.1
flatbuffers==24.3.25
gast==0.6.0
google-pasta==0.2.0
greenlet==3.1.1
grpcio==1.66.2
h5py==3.12.1
idna==3.10
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keras==3.6.0
kombu==5.4.2
libclang==18.1.1
Mako==1.3.5
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.1
mdurl==0.1.2
ml-dtypes==0.4.1
namex==0.0.8
numpy==1.26.4
opt_einsum==3.4.0
optree==0.13.0
packaging==24.1
pandas==2.2.3
prompt_toolkit==3.0.48
protobuf==4.25.5
psycopg2-binary==2.9.9
Pygments==2.18.0
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
redis==5.1.1
referencing==0.35.1
requests==2.32.3
rich==13.9.2
rpds-py==0.20.0
scikit-learn==1.5.2
scipy==1.14.1
six==1.16.0
soupsieve==2.6
SQLAlchemy==2.0.35
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-intel==2.17.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.5.0
threadpoolctl==3.5.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.0.4
wrapt==1.16.0
openai>=1.0.0
python-dotenv
requests
flask
Flask-Mail
Flask-Login
Flask-Limiter